Partially update to the kickoff release of Phoenix
They made some pretty major changes to the Swerve API, so I have to update some of the code for it to be ideal. It should work for now (except for advantagekit crashing since kickoff release isn't out yet), but I'll try to update the code to fully match tommorow.

Signed-off-by: Ashton A <uhdbits@duck.com>